+++ cli/cmd/rocks.go
@@ -18,7 +18,8 @@ func NewRocksCmd() *cobra.Command {
 		DisableFlagParsing: true,
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.Name(), &modulesInfo, internalRocksModule, args)
+			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
+				internalRocksModule, args)
 			if err != nil {
 				log.Fatalf(err.Error())
 			}