+++ cli/cmd/run.go
@@ -52,7 +52,7 @@ are passed after '--'.
 `,
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.Name(), &modulesInfo, internalRunModule, args)
+			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo, internalRunModule, args)
 			if err != nil {
 				log.Fatalf(err.Error())
 			}