+++ cli/cmd/start.go
@@ -26,7 +26,8 @@ func NewStartCmd() *cobra.Command {
 		Short: "Start tarantool instance(s)",
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdCtx.CommandName = cmd.Name()
-			err := modules.RunCmd(&cmdCtx, cmd.Name(), &modulesInfo, internalStartModule, args)
+			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
+				internalStartModule, args)
 			if err != nil {
 				log.Fatalf(err.Error())
 			}